api/python/requirements.txt
tornado>=4.3
pyyaml>=3.11
sqlalchemy
yagmail
